Revert "CONSOLE-4272: report csp violations to telemtry"
