+++ frontend/public/locales/en/public.json
@@ -53,6 +53,8 @@
   "Details": "Details",
   "Schema": "Schema",
   "Instances": "Instances",
+  "Content Security Policy violation in Console plugin": "Content Security Policy violation in Console plugin",
+  "{{pluginName}} might have violated the Console Content Security Policy. Refer to the browser's console logs for details.": "{{pluginName}} might have violated the Console Content Security Policy. Refer to the browser's console logs for details.",
   "Web console update is available": "Web console update is available",
   "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.": "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.",
   "Refresh web console": "Refresh web console",
@@ -1941,8 +1943,6 @@
   "Select a log file above": "Select a log file above",
   "Filter by unit": "Filter by unit",
   "Unit": "Unit",
-  "Content Security Policy violation in Console plugin": "Content Security Policy violation in Console plugin",
-  "{{pluginName}} might have violated the Console Content Security Policy. Refer to the browser's console logs for details.": "{{pluginName}} might have violated the Console Content Security Policy. Refer to the browser's console logs for details.",
   "Loading {{title}} status": "Loading {{title}} status",
   "graph timespan": "graph timespan",
   "Reset zoom": "Reset zoom",
